nicify things
